backend/requirements.txt
Django>=4.2,<5
djangorestframework
djangorestframework-simplejwt
django-filter
channels
daphne
psycopg2-binary
redis
celery
django-cors-headers
drf-spectacular
boto3
django-storages
python-dotenv
django-jazzmin
whitenoise
Pillow
django-mptt
